% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{align_phase}
\alias{align_phase}
\title{Align two phase vectors at 0.}
\usage{
align_phase(phi, target, center = TRUE)
}
\arguments{
\item{phi}{phase angles to be shifted (radian).}

\item{target}{phase angles of the same elements, where 0 is the anchor.}

\item{center}{re-center the shifted phases in -pi:pi.}
}
\value{
the shifted \code{phi}.
}
\description{
Shifts \code{phi} by its value where \code{target} is 0 (linear
interpolation).
}
