% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{circ.axis}
\alias{circ.axis}
\title{generate circular plot axis in radian}
\usage{
circ.axis(
  x,
  at = c(-2 * pi, -pi, -pi/2, 0, pi/2, pi, 1.5 * pi, 2 * pi, 2.5 * pi, 3 * pi),
  labels = expression(-2 * pi, -pi, -pi/2, 0, pi/2, pi, "", 2 * pi, "", 3 * pi),
  ...
)
}
\arguments{
\item{x}{side(s) of the plot, as in \code{\link[graphics]{axis}}.}

\item{at, labels}{tick positions and labels, multiples of pi/2.}

\item{...}{further arguments to \code{\link[graphics]{axis}}.}
}
\description{
generate circular plot axis in radian
}
