% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{evaluate_order}
\alias{evaluate_order}
\title{Calculate the difference between input and calculated state order.}
\usage{
evaluate_order(phases)
}
\arguments{
\item{phases}{phases object as returned by \link{get_pseudophase}.}
}
\value{
the distance, see \link{state_order_distance}.
}
\description{
The differences are calculated as Levenshtein distances, using a
simple circular extension of the base R \link{adist} function.
}
