% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_degradation}
\alias{get_degradation}
\title{Fit a transcript degradation rate from relative amplitude and
oscillation parameters, using the \code{\link[stats]{uniroot}}  function}
\usage{
get_degradation(
  a,
  R,
  Rmin,
  phi,
  tau,
  mu,
  model,
  lower = 1e-06,
  upper = 10000,
  tol = 1e-09,
  verb = 0,
  ...
)
}
\arguments{
\item{a}{relative abundance amplitude, (max(x)-min(x))/mean(x).}

\item{R}{mean abundance, required for model with basal transcription.}

\item{Rmin}{minimal abundance, required for model with basal transcription.}

\item{phi}{duty cycle.}

\item{tau}{period.}

\item{mu}{growth rate, used only for correction (gamma=growth+degradation).}

\item{model}{model name; a function \code{root_<model>} must exist.}

\item{lower}{the lower end point of the interval to be  searched by \code{\link[stats]{uniroot}}.}

\item{upper}{the upper end point of the interval to be  searched by \code{\link[stats]{uniroot}}.}

\item{tol}{error tolerance of the \code{\link[stats]{uniroot}}  call.}

\item{verb}{verbosity; with 0, errors of the root finding are silent.}

\item{...}{unused.}
}
\value{
the degradation rate \code{gamma - mu} (or \code{gamma} if
\code{mu} is missing or \code{NA}), \code{NA} if no root was found.
}
\description{
Fit a transcript degradation rate from relative amplitude and
oscillation parameters, using the \code{\link[stats]{uniroot}}  function
}
\keyword{internal}
