% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts_utils.R, R/phases.R
\name{phase_align}
\alias{phase_align}
\title{align to phase vectors at 0}
\usage{
phase_align(phi, target, shift = TRUE)

phase_align(phi, target, shift = TRUE)
}
\description{
align to phase vectors at 0

align to phase vectors at 0
}
