% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{phase_rank}
\alias{phase_rank}
\title{Returns the rank of a phase angle in circular coordinates.}
\usage{
phase_rank(theta, align = TRUE, center = TRUE)
}
\arguments{
\item{align}{align the original phase angle and the rank phase
angle at phase 0.}

\item{center}{center the aligned phase angle such that all phases are
in -pi:pi, using \link{align_phase}.}
}
\description{
Returns the rank of a phase angle in circular coordinates.
}
