% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{phase_segments}
\alias{phase_segments}
\title{Classify phases into segments.}
\usage{
phase_segments(phi, breaks, lim = c(-pi, pi))
}
\arguments{
\item{phi}{phase angles.}

\item{breaks}{segment borders.}

\item{lim}{range of the phase angles.}
}
\value{
integer segment number for each phase.
}
\description{
The first and last segment, both ends of the circle, are fused.
}
