% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{phaseHist}
\alias{phaseHist}
\title{Plot a phase histogram by cell classes.}
\usage{
phaseHist(phase, cls, cls.srt, cls.col, sid = "", leg.nrow = 1)
}
\arguments{
\item{phase}{phase of each cell.}

\item{cls}{class of each cell.}

\item{cls.srt}{classes, in legend order.}

\item{cls.col}{colors (hex, without alpha), named by class.}

\item{sid}{plot title.}

\item{leg.nrow}{number of legend rows.}
}
\description{
Plot a phase histogram by cell classes.
}
