% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGOI}
\alias{plotGOI}
\title{Plot genes of interest time series}
\usage{
plotGOI(
  phases,
  counts,
  goi,
  names,
  col,
  win = 0.05,
  xlab = expression(pseudophase * phi),
  ylab = "mov. avg. of counts/total"
)
}
\arguments{
\item{phases}{phases.}

\item{counts}{count table, genes x cells.}

\item{goi}{genes of interest, rows of \code{counts}.}

\item{names}{legend names.}

\item{col}{colors.}

\item{win}{fraction of cells for the moving average.}

\item{xlab, ylab}{axis labels.}
}
\description{
Currently stops: needs to be updated to the phases object.
}
