% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{plotSegments}
\alias{plotSegments}
\title{Plot the curve sketching of theta=f(phi).}
\usage{
plotSegments(
  phases,
  difference = FALSE,
  center = TRUE,
  method = "shoulder",
  show.cohorts = TRUE,
  axis4 = FALSE,
  legend = FALSE,
  show.max = TRUE
)
}
\arguments{
\item{phases}{phases object as returned by \link{get_pseudophase}.}

\item{difference}{plot the difference theta-phi instead of
theta. This emphasizes the noise in the original data and the
smoothing effects.}

\item{center}{remove the jump in theta and center it in -pi:pi.}

\item{method}{the segmentation to show, by its name in
\link{get_segments}.}

\item{show.cohorts}{label the cohort phases on the top axis.}

\item{axis4}{add an axis for the derivative on the right.}

\item{legend}{add a legend for the derivative.}

\item{show.max}{unused.}
}
\description{
Plot the curve sketching of theta=f(phi).
}
