% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotStates}
\alias{plotStates}
\title{Plot cohort state time series.}
\usage{
plotStates(
  phase,
  states,
  cls.srt,
  cls.col,
  ncells,
  ordered = FALSE,
  row.center = FALSE,
  scale = FALSE,
  center = FALSE,
  win = 0.01,
  circular = TRUE,
  ma = TRUE,
  log = FALSE,
  norm = FALSE,
  lines = TRUE,
  sid = "",
  legend = TRUE,
  leg.nrow = 1,
  leg.pos = "topright",
  xtype = "phase",
  xlab = expression(phase ~ phi),
  xlim = c(-pi, pi),
  lwd = 2,
  axes = TRUE,
  ylab,
  ylim,
  verb = 0,
  ...
)
}
\arguments{
\item{phase}{phase of each cell (column of \code{states}).}

\item{states}{a cohort expression state table, see \link{get_states}.}

\item{cls.srt}{states to plot, in this order.}

\item{cls.col}{colors, named by state.}

\item{ncells}{width of the moving average in cells; default:
\code{win} times the number of cells.}

\item{ordered}{\code{phase} and the columns of \code{states} are
ordered already.}

\item{row.center, scale, center}{processing as in \link{get_pseudophase}.}

\item{win}{fraction of cells for the moving average.}

\item{circular}{circular moving average.}

\item{ma}{plot the moving average.}

\item{log}{unused.}

\item{norm}{plot the log2 ratio to the state mean.}

\item{lines}{plot the raw data.}

\item{sid}{plot label, see \code{segmenTools::figlabel}.}

\item{legend}{add a legend.}

\item{leg.nrow}{number of legend rows.}

\item{leg.pos}{legend position.}

\item{xtype}{\code{"phase"} or \code{"angle"}: circular x axis.}

\item{xlab, ylab, xlim, ylim, lwd, axes}{graphics parameters.}

\item{verb}{verbosity.}

\item{...}{further arguments to \code{\link[graphics]{plot}}.}
}
\description{
Each state along the phase of the cells, as raw lines and a moving
average.
}
