% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcycle-package.R
\docType{package}
\name{rcycle-package}
\alias{rcycle}
\alias{rcycle-package}
\title{rcycle: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Rainer Machne \email{rainer.machne@gmail.com} (\href{https://orcid.org/ 0000-0002-1274-5099 }{ORCID})

}
\keyword{internal}
