% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{remove_jumps}
\alias{remove_jumps}
\title{Remove a jump in phases by shift all phases.}
\usage{
remove_jumps(phi, idx, center = TRUE, verb = 1)
}
\arguments{
\item{phi}{ordered list of phases}

\item{idx}{position of the jump; default: \link{detect_jumps}.}

\item{center}{shift the result into -pi:pi.}

\item{verb}{verbosity.}
}
\value{
the phases without the jump.
}
\description{
Remove a jump in phases by shift all phases.
}
