% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{revert}
\alias{revert}
\title{Revert all phases if order is wrong.}
\usage{
revert(phases, srt, force = FALSE, verb = 1)
}
\arguments{
\item{phases}{phases object as returned by \link{get_pseudophase}.}

\item{srt}{optional names of the states (cohorts) to use for the order
test.}

\item{force}{revert without testing.}

\item{verb}{verbosity.}
}
\value{
the phases object.
}
\description{
This function compares the calculated order of states to the input
order of the state matrix. If the reverse order better matches the
input order, all phases are reverted.
}
