% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts_utils.R, R/phases.R
\name{revert_phases}
\alias{revert_phases}
\title{Establish whether the phases should be reverted wrt state order.}
\usage{
revert_phases(state, phase, window = 0.05, verb = 1)

revert_phases(state, phase, window = 0.05, verb = 1)
}
\description{
Establish whether the phases should be reverted wrt state order.

Establish whether the phases should be reverted wrt state order.
}
