% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{shift}
\alias{shift}
\title{Shift all phases in a phase object.}
\usage{
shift(phases, dphi, align = FALSE, center = FALSE, verb = 1)
}
\arguments{
\item{phases}{phases object as returned by \link{get_pseudophase}.}

\item{dphi}{phase angle by which the phase phi in the phases object
will be shifted.}

\item{align}{align all \code{theta*} angles to the shifted \code{phi}
at 0, with \link{align_phase}.}

\item{center}{re-center the aligned \code{theta*} in -pi:pi.}

\item{verb}{verbosity.}
}
\value{
the phases object.
}
\description{
Shifts \code{phi} of every element that has one, re-orders, and
optionally aligns the \code{theta*} angles at the new 0.
}
