% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts_utils.R, R/phases.R
\name{shift_phases}
\alias{shift_phases}
\title{Shift all phases in a phase object.}
\usage{
shift_phases(phases, dph, shift = TRUE, verb = 0)

shift_phases(phases, dph, shift = TRUE, verb = 0)
}
\description{
Shift all phases in a phase object.

Shift all phases in a phase object.
}
