% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{arrows.phases}
\alias{arrows.phases}
\title{Add arrows for segments in the phases object.}
\usage{
arrows.phases(
  x,
  y,
  types = "shoulder",
  phase = "phi",
  y0,
  dy,
  col,
  labels,
  labels.top,
  lxpd = par("xpd"),
  pos = NULL,
  ticks = FALSE,
  verb = 0,
  ...
)
}
\description{
Add arrows for segments in the phases object.
}
