% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate phase to a period.}
\usage{
calibrate(phases, period, phase = "phi")
}
\arguments{
\item{phases}{phases object as returned by \link{get_pseudophase}.}

\item{period}{a period of the transcription cycle, in units of
time.}

\item{phase}{the phase ID in the phases object to use for calibration
to the transcription cycle period.}
}
\description{
Calibrate phase to a period.
}
