% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{center}
\alias{center}
\title{Center phases at anchor points.}
\usage{
center(phases, method = "slope", params = list(spar = 0.001), ...)
}
\description{
High-level interface for the \link{shift} function that allows to
specify how to select the anchor point.
}
