% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{clustering2cohorts}
\alias{clustering2cohorts}
\title{convert a gene clusterng vector to a gene cohort list}
\usage{
clustering2cohorts(cls, cls.srt)
}
\description{
convert a gene clusterng vector to a gene cohort list
}
