% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{cohort2clustering}
\alias{cohort2clustering}
\title{Convert a gene cohort list or matrix to a vector of unique gene classes.}
\usage{
cohort2clustering(cohorts, n, na = "na")
}
\description{
Convert a gene cohort list or matrix to a vector of unique gene classes.
}
