% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{cohort2clustering}
\alias{cohort2clustering}
\title{convert a gene cohort list or matrix to a clustering vector}
\usage{
cohort2clustering(cohorts, n, na = "na")
}
\description{
convert a gene cohort list or matrix to a clustering vector
}
