% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts_utils.R, R/phases.R
\name{detect_jumps}
\alias{detect_jumps}
\title{Return the position of a jump}
\usage{
detect_jumps(py, max = -pi, verb = 0)

detect_jumps(py, max = -pi, verb = 0)
}
\description{
Return the position of a jump

Return the position of a jump
}
