% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts_utils.R, R/phases.R
\name{get_classes}
\alias{get_classes}
\title{Classify cells by maximal state log2 fold change.}
\usage{
get_classes(state)

get_classes(state)
}
\description{
Classify cells by maximal state log2 fold change.

Classify cells by maximal state log2 fold change.
}
