% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{get_cohorts}
\alias{get_cohorts}
\title{Convert a gene cohort list to a boolean or size-normalized cohort matrix}
\usage{
get_cohorts(cohorts, n, normalize = FALSE)
}
\arguments{
\item{cohorts}{a list of integer vectors, each providing the row indices
of the genes in the count table}

\item{n}{the number of rows (genes) in the count table, if not provided
the maximal index will be used}

\item{normalize}{divide each boolean entry (1) by the cohort size}
}
\description{
Convert a gene cohort list to a boolean or size-normalized cohort matrix
}
