% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_degradation}
\alias{get_degradation}
\title{Fit a transcript degradation rate from relative amplitude and
oscillation parameters, using the \code{\link[stats]{uniroot}}  function}
\usage{
get_degradation(
  a,
  R,
  Rmin,
  phi,
  tau,
  mu,
  model,
  lower = 1e-06,
  upper = 10000,
  tol = 1e-09,
  verb = 0,
  ...
)
}
\arguments{
\item{a}{relative abundance amplitude (max(x)-min(x)/mean(x)).}

\item{R}{mean abundance, required for model with basal transcription.}

\item{Rmin}{minimal abundance, required for model with basal transcription.}

\item{phi}{duty cycle.}

\item{tau}{period.}

\item{mu}{growth rate, used only for correction (gamma=growth+degradation).}

\item{model}{}

\item{lower}{the lower end point of the interval to be  searched by \code{\link[stats]{uniroot}}.}

\item{upper}{the upper end point of the interval to be  searched by \code{\link[stats]{uniroot}}.}

\item{tol}{error tolerance of the \code{\link[stats]{uniroot}}  call.}

\item{...}{further parameters to \code{\link[stats]{uniroot}}}

\item{k}{transcription rate, required for model with basal transcription.}
}
\description{
Fit a transcript degradation rate from relative amplitude and
oscillation parameters, using the \code{\link[stats]{uniroot}}  function
}
