% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{get_pseudophase}
\alias{get_pseudophase}
\title{get the pseudophase from a cohort state matrix}
\usage{
get_pseudophase(
  states,
  center,
  revert.phase = FALSE,
  window = 0.05,
  segments = FALSE,
  spar = 0.001,
  classify = FALSE,
  validate = FALSE,
  add.loadings = TRUE,
  row.center = TRUE,
  verb = 1
)
}
\arguments{
\item{state}{a cohort expression state table, with cohort mean
counts as rows and cells as columns.}
}
\description{
get the pseudophase from a cohort state matrix
}
