% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts_utils.R, R/phases.R
\name{get_pseudophase}
\alias{get_pseudophase}
\title{Get the pseudophase from a cohort state matrix.}
\usage{
get_pseudophase(
  state,
  center,
  revert.phase = FALSE,
  window = 0.05,
  segments = FALSE,
  spar = 0.001,
  classify = FALSE,
  validate = FALSE,
  add.loadings = TRUE,
  unit = FALSE,
  row.center = TRUE,
  verb = 1
)

get_pseudophase(
  state,
  center,
  revert.phase = FALSE,
  window = 0.05,
  segments = FALSE,
  spar = 0.001,
  classify = FALSE,
  validate = FALSE,
  add.loadings = TRUE,
  unit = FALSE,
  row.center = TRUE,
  verb = 1
)
}
\arguments{
\item{state}{a cohort expression state table, with cohort mean
counts as rows and cells as columns.}
}
\description{
Get the pseudophase from a cohort state matrix.

get the pseudophase from a cohort state matrix
}
