% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{get_pseudophase}
\alias{get_pseudophase}
\title{get the pseudophase from a cohort state matrix}
\usage{
get_pseudophase(
  states,
  revert.phase = FALSE,
  segments = FALSE,
  spar = 0.001,
  classify = FALSE,
  validate = FALSE,
  row.center = TRUE,
  verb = 1
)
}
\arguments{
\item{states}{a cohort expression state table, with cohort mean
counts as rows and cells as columns, as provided by
\link{get_states}.}

\item{verb}{output verbosity level.}
}
\description{
get the pseudophase from a cohort state matrix
}
