% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_ramp}
\alias{get_ramp}
\title{Calculate abundance amplitudes from rates and times.}
\usage{
get_ramp(
  gamma,
  dr,
  mu,
  phi,
  tau,
  relative = TRUE,
  k,
  k0,
  force.relative = FALSE,
  use.coth = FALSE,
  model = c("k", "dr", "k_dr", "k_dr_k0"),
  ...
)
}
\description{
Calculate abundance amplitudes from rates and times.
}
