% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_rmean}
\alias{get_rmean}
\title{Calculate mean abundance from rates and times.}
\usage{
get_rmean(
  k,
  gamma,
  k0,
  dr,
  mu,
  phi,
  tau,
  model = c("k", "dr", "k_dr", "k_dr_k0"),
  use.coth = TRUE
)
}
\description{
Calculate mean abundance from rates and times.
}
