% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{get_state_order}
\alias{get_state_order}
\title{Get the order of states (rows of the cohort state matrix).}
\usage{
get_state_order(
  states,
  window = 0.05,
  sides = 2,
  circular = TRUE,
  names = FALSE
)
}
\arguments{
\item{states}{an ORDERED cohort expression state table, with cohort
mean counts as rows and cells as columns, as provided by
\link{get_states}. The columns must be ordered for the moving
average calculation.}

\item{window}{fraction of cells to be used for the moving average
of states.}

\item{circular}{treat the data as circular in the moving average
calculation.}
}
\description{
NOTE: cells (columns of the state matrix) must be ordered!
}
