% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{get_state_order}
\alias{get_state_order}
\title{Get the order of states (rows of the cohort state matrix).}
\usage{
get_state_order(
  states,
  window = 0.05,
  sides = 2,
  circular = TRUE,
  names = FALSE
)
}
\description{
Get the order of states (rows of the cohort state matrix).
}
