% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_transcription}
\alias{get_transcription}
\title{Calculate transcription rate from average abundance and degradation rates.
from base equation for average abundance
TODO: map fluorescence to a rough transcript/cell number}
\usage{
get_transcription(
  R = NA,
  A = NA,
  phi = NA,
  dr = NA,
  mu = NA,
  gamma = NA,
  tau = NA,
  model = c("k", "dr", "k_dr", "k_dr_k0")
)
}
\arguments{
\item{R}{average abundance}

\item{A}{absolute amplitude, for biphasic models}
}
\description{
Calculate transcription rate from average abundance and degradation rates.
from base equation for average abundance
TODO: map fluorescence to a rough transcript/cell number
}
