% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillation.R
\name{get_wavelet}
\alias{get_wavelet}
\title{Wavelet analysis of counts along a circular pseudophase.}
\usage{
get_wavelet(phi, counts, ID, verb = 1, ...)
}
\arguments{
\item{phi}{a circular pseudophase assignement for the counts vector
or matrix.}

\item{counts}{a vector or matrix of counts of the same dimension
and order as the pseudophase phi.}

\item{ID}{if counts is a matrix, ID specifies the row number or row
name for which the analysis is to be performed.}

\item{verb}{verbosity level, set to >2 to also get verbose output
from the wavelet function.}

\item{...}{arguments to \code{\link[WavletComp]{analyze.wavelet}}.}
}
\description{
Wavelet analysis of counts along a circular pseudophase.
}
