% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillation.R
\name{morlet_wavelets}
\alias{morlet_wavelets}
\title{Morlet wavelet function (via chatGPT)}
\usage{
morlet_wavelets(t, f0 = 6, s0 = 1, dj = 1/12, j = 0)
}
\arguments{
\item{t}{time vector in sec}

\item{central}{frequency in Hz}
}
\description{
Morlet wavelet function (via chatGPT)
}
