% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize count table by total counts per cell}
\usage{
normalize_counts(counts, check = TRUE)
}
\arguments{
\item{count}{table with genes in rows and cells in columns}
}
\description{
Normalize count table by total counts per cell
}
