% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts_utils.R, R/phases.R
\name{phase_rank}
\alias{phase_rank}
\title{Returns the rank of a phase angle in circular coordinates (radian).}
\usage{
phase_rank(phi, align = TRUE, shift = TRUE)

phase_rank(phi, align = TRUE, shift = TRUE)
}
\arguments{
\item{align}{align the original phase angle and the rank phase
angle at phase 0.}

\item{shift}{shift the aligned phase angle such that all phases are
in -pi:pi, using \link{phase_align}.}
}
\description{
Returns the rank of a phase angle in circular coordinates (radian).
}
