% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGOI}
\alias{plotGOI}
\title{Plot genes of interest time series}
\usage{
plotGOI(
  phases,
  counts,
  goi,
  names,
  col,
  win = 0.05,
  xlab = expression(pseudophase * phi),
  ylab = "mov. avg. of counts/total"
)
}
\description{
Plot genes of interest time series
}
