% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPC}
\alias{plotPC}
\title{Plot PCA-based circle and state vectors.}
\usage{
plotPC(
  phases,
  x = 1,
  y = 2,
  col,
  expand = TRUE,
  eigen.axis = TRUE,
  cohorts = TRUE,
  ccol,
  txt.cex = 1,
  ...
)
}
\description{
Plot PCA-based circle and state vectors.
}
