% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPC}
\alias{plotPC}
\title{Plot PCA-based circle and state vectors.}
\usage{
plotPC(
  phases,
  x = 1,
  y = 2,
  z,
  z.q = c(0.05, 0.95),
  z.legend = FALSE,
  vectors = TRUE,
  vlines = FALSE,
  varrows = FALSE,
  vlabels = FALSE,
  vlabels.top = Inf,
  colf = NULL,
  col = NULL,
  lwd = 1,
  pch = 19,
  cex = 0.5,
  vaxis = vectors,
  scores = TRUE,
  slines = FALSE,
  sarrows = FALSE,
  slabels = FALSE,
  slabels.top,
  scolf = NULL,
  scol = NULL,
  slwd = 1,
  spch = 1,
  scex = 1,
  saxis = scores,
  txt.cex = 1,
  scale = 1,
  pc.biplot = FALSE,
  xlim,
  ylim,
  expand = 1,
  arcsinh = FALSE,
  zero.axis = FALSE,
  zero.axis.label = zero.axis,
  pc.arrows = FALSE,
  pc.lwd = 1,
  pc.scale = 1,
  show.var = TRUE,
  ...
)
}
\arguments{
\item{phases}{PCA object, as returned from stats::prcomp or the
annotated version via rcycle::pseudophase.}

\item{x, y}{PC components to plot on x and y axes.}

\item{z}{use the zth PC component for coloring between quantiles
given in z.q}

\item{vectors}{plot eigenvectors.}

\item{vlines}{connect eigenvector tips by a line, useful for
circular data.}

\item{varrows}{draw origin-based arrows for eigenvectors (as in a biplot).}

\item{scores}{plot transformed data.}

\item{slines}{connect transformed data by a line, useful for
circular data.}

\item{sarrows}{draw origin-based arrows for transformed data.}

\item{scale}{scale parameter, if provided it converts the plot to a
true biplot}
}
\description{
NOTE: scale = 1, pc.biplot = FALSE is equivalent to the defaults
of R's biplot.prcomp.
}
