% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPC}
\alias{plotPC}
\title{Plot PCA-based circle and state vectors.}
\usage{
plotPC(
  phases,
  x = 1,
  y = 2,
  z,
  z.q = c(0.05, 0.95),
  z.legend = FALSE,
  vectors = TRUE,
  vlines = FALSE,
  varrows = FALSE,
  vlabels = FALSE,
  colf = NULL,
  col = NULL,
  pch = 19,
  cex = 0.5,
  vaxis = vectors,
  scores = TRUE,
  slines = FALSE,
  sarrows = FALSE,
  slabels = FALSE,
  scolf = NULL,
  scol = NULL,
  spch = 1,
  scex = 1,
  saxis = scores,
  txt.cex = 1,
  scale = 0,
  xlim,
  ylim,
  expand = 1,
  pc.biplot = FALSE,
  arcsinh = FALSE,
  zero.axis = FALSE,
  show.var = TRUE,
  ...
)
}
\arguments{
\item{z}{use the zth PC component for coloring between quantiles given in z.q}

\item{scale}{scale parameter, if provided it converts the plot to a true biplot}
}
\description{
Plot PCA-based circle and state vectors.
}
