% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{plotSegments}
\alias{plotSegments}
\title{Plot the curve sketching of theta=f(phi).}
\usage{
plotSegments(
  phases,
  difference = FALSE,
  center = TRUE,
  method = "shoulder",
  show.cohorts = TRUE
)
}
\arguments{
\item{difference}{plot the difference theta-phi instead of
theta. This emphasizes the noise in the original data and the
smoothing effects.}
}
\description{
Plot the curve sketching of theta=f(phi).
}
