% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotStates}
\alias{plotStates}
\title{Plot cohort state time series.}
\usage{
plotStates(
  phase,
  states,
  cls.srt,
  cls.col,
  ncells,
  ordered = FALSE,
  win = 0.01,
  circular = TRUE,
  ma = TRUE,
  log = FALSE,
  norm = FALSE,
  lines = TRUE,
  sid = "",
  legend = TRUE,
  leg.nrow = 1,
  leg.pos = "topright",
  xtype = "phase",
  xlab = expression(phase ~ phi),
  xlim = c(-pi, pi),
  lwd = 2,
  axes = TRUE,
  ylab,
  ylim,
  verb = 0,
  ...
)
}
\description{
Plot cohort state time series.
}
