% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotStates}
\alias{plotStates}
\title{Plot cohort state time series.}
\usage{
plotStates(
  phases,
  states,
  xtype = "phase",
  cls.srt,
  cls.col,
  win = 0.01,
  norm = FALSE,
  lines = TRUE,
  sid = "",
  legend = TRUE,
  leg.nrow = 1,
  xlab,
  ylab,
  ...
)
}
\description{
Plot cohort state time series.
}
