% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pw_fourier}
\alias{pw_fourier}
\title{alpha controls the smoothing by exponentially damping higher harmonics; alpha=0: no smoothing; alpha<=0.5: mild smoothing; alpha>0.5: noticable smoothing; alpha>2: towards sinus}
\usage{
pw_fourier(
  t = 0,
  k,
  phi,
  tosc,
  N = 10000,
  shift = 0,
  theta = 0,
  start.on = FALSE,
  alpha = 0
)
}
\description{
t <- seq(0,10, length=100)
plot(t, pwf(t=t, tosc=1.5, k=1, phi=.4), type='l', col=2)
lines(t, pwf(t=t, tosc=1.5, k=1, phi=.6, theta=pi), type='l', col=4)
}
