% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pwmode_k_dr}
\alias{pwmode_k_dr}
\title{ODE of pulse width-modulated transcription and anti-phasic
degradation.}
\usage{
pwmode_k_dr(time, state, parameters, hocf)
}
\description{
ODE of pulse width-modulated transcription and anti-phasic
degradation.
}
