% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pwmode_k_dr_k0}
\alias{pwmode_k_dr_k0}
\title{ODE of pulse width-modulated transcription and anti-phasic degradation
and basal transcription.}
\usage{
pwmode_k_dr_k0(time, state, parameters, hocf)
}
\description{
ODE of pulse width-modulated transcription and anti-phasic degradation
and basal transcription.
}
