% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{remove_jumps}
\alias{remove_jumps}
\title{Remove a jump in phases by shift all phases.}
\usage{
remove_jumps(phi, shift = TRUE, verb = 1)
}
\arguments{
\item{phi}{ordered list of phases}
}
\description{
Remove a jump in phases by shift all phases.
}
