% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{revert_phase}
\alias{revert_phase}
\title{Establish whether the phases should be reverted wrt state order.}
\usage{
revert_phase(phi, states, ...)
}
\description{
Establish whether the phases should be reverted wrt state order.
}
