% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{revert_phase_state}
\alias{revert_phase_state}
\title{Establish whether the phases should be reverted wrt state order.}
\usage{
revert_phase_state(states, phi, window = 0.05)
}
\description{
Establish whether the phases should be reverted wrt state order.
}
