% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{segment_state}
\alias{segment_state}
\title{Classify segments by their overlap with cell classes.}
\usage{
segment_state(
  phases,
  segment = "inflection",
  class = "class",
  col,
  plot = FALSE
)
}
\description{
Classify segments by their overlap with cell classes.
}
