% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{shift}
\alias{shift}
\title{Shift all phases in a phase object.}
\usage{
shift(phases, dphi, align = FALSE, center = FALSE, verb = 1)
}
\arguments{
\item{phases}{phases object as returned by \link{get_pseudophase}.}

\item{dphi}{phase angle by which the phase phi in the phases object
will be shifted.}

\item{center}{re-center the shifted theta (align is true)
between -pi:pi.}

\item{verb}{output verbosity level.}
}
\description{
Shift all phases in a phase object.
}
