% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{shift}
\alias{shift}
\title{Shift all phases and angles in a phase object.}
\usage{
shift(phases, dphi, align = FALSE, center = FALSE, verb = 1)
}
\description{
Shift all phases and angles in a phase object.
}
