% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.R
\name{state_order_distance}
\alias{state_order_distance}
\title{calculate Levenshtein distance between state/test and a reference order}
\usage{
state_order_distance(state, test, reference, ...)
}
\description{
calculate Levenshtein distance between state/test and a reference order
}
