% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts_utils.R, R/phases.R
\name{state_phase}
\alias{state_phase}
\title{Get the phase of a cohort state.}
\usage{
state_phase(state, phase, center, window = 0.05, verb = 0)

state_phase(state, phase, center, window = 0.05, verb = 0)
}
\description{
Get the phase of a cohort state.

Get the phase of a cohort state.
}
